% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models-utils.R
\name{build_nested_tibble_from_safely_output}
\alias{build_nested_tibble_from_safely_output}
\title{Create a tibble from the output of linear_model_safely(),
logistic_model_safely(), cox_model_safely()}
\usage{
build_nested_tibble_from_safely_output(z)
}
\description{
Create a tibble from the output of linear_model_safely(),
logistic_model_safely(), cox_model_safely()
}
\keyword{internal}
