% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence-rates.R
\name{calc_rates}
\alias{calc_rates}
\title{Calculate rates and their standard errors using various methods}
\usage{
calc_rates(data, per_pyar, method = c("norm", "gamma"))
}
\arguments{
\item{data}{A data frame.}

\item{per_pyar}{Crude and standardised incidence rates are given per \code{per_pyar} person-years at risk. (Default: 1000)}

\item{method}{Method to calculate confidence intervals, one of "norm" or "gamma". Default "norm" is Normal approximation.}
}
\value{
A data frame
}
\description{
Calculate rates and their standard errors using various methods
}
\keyword{internal}
