% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckbtools-package.R
\docType{package}
\name{ckbtools-package}
\alias{ckbtools}
\alias{ckbtools-package}
\title{ckbtools: Tools for Analyses in Epidemiology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A selection of tools for analyses in epidemiology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ckbiobank.github.io/ckbtools}
}

}
\author{
\strong{Maintainer}: Neil Wright \email{neil.wright@ndph.ox.ac.uk}

Authors:
\itemize{
  \item Neil Wright \email{neil.wright@ndph.ox.ac.uk}
  \item Derrick Bennett
  \item Jonathan Clarke
  \item Charlotte Clarke
  \item Andri Iona
  \item Christiana Kartsonaki
  \item Maryam Rahmati
  \item China Kadoorie Biobank Collaborative Group
}

Other contributors:
\itemize{
  \item University of Oxford [copyright holder]
}

}
\keyword{internal}
