% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbtools_ascertainments}
\alias{ckbtools_ascertainments}
\title{Synthetic CKB ascertainment data for demoing functions}
\format{
An object of class \code{data.frame} with 25000 rows and 3 columns.
}
\usage{
ckbtools_ascertainments
}
\description{
Synthetic CKB ascertainment data for demoing functions
}
\keyword{datasets}
