% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbtools_baseline}
\alias{ckbtools_baseline}
\title{Synthetic CKB baseline data for demoing functions}
\format{
An object of class \code{data.frame} with 25000 rows and 26 columns.
}
\usage{
ckbtools_baseline
}
\description{
Synthetic CKB baseline data for demoing functions
}
\keyword{datasets}
