% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbtools_endpoint_definitions}
\alias{ckbtools_endpoint_definitions}
\title{Synthetic CKB endpoint definitions data for demoing functions}
\format{
An object of class \code{data.frame} with 2 rows and 2 columns.
}
\usage{
ckbtools_endpoint_definitions
}
\description{
Synthetic CKB endpoint definitions data for demoing functions
}
\keyword{datasets}
