% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbtools_endpoints}
\alias{ckbtools_endpoints}
\title{Synthetic CKB endpoint data for demoing functions}
\format{
An object of class \code{data.frame} with 25000 rows and 5 columns.
}
\usage{
ckbtools_endpoints
}
\description{
Synthetic CKB endpoint data for demoing functions
}
\keyword{datasets}
