% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbtools_participant_data}
\alias{ckbtools_participant_data}
\title{Synthetic CKB participant data for demoing functions}
\format{
An object of class \code{data.frame} with 25000 rows and 47 columns.
}
\usage{
ckbtools_participant_data
}
\description{
Synthetic CKB participant data for demoing functions
}
\keyword{datasets}
