% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbtools_raw_olink_protein_data}
\alias{ckbtools_raw_olink_protein_data}
\title{Synthetic CKB Olink proteomics data for demoing functions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48000 rows and 11 columns.
}
\usage{
ckbtools_raw_olink_protein_data
}
\description{
Synthetic CKB Olink proteomics data for demoing functions
}
\keyword{datasets}
