% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ckbtools_raw_somascan_meta_data}
\alias{ckbtools_raw_somascan_meta_data}
\title{Synthetic CKB SomaScan proteomics data for demoing functions}
\format{
An object of class \code{data.frame} with 48 rows and 7 columns.
}
\usage{
ckbtools_raw_somascan_meta_data
}
\description{
Synthetic CKB SomaScan proteomics data for demoing functions
}
\keyword{datasets}
