% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models.R
\name{cox_model}
\alias{cox_model}
\title{Estimate the association between an exposure and incident disease}
\usage{
cox_model(
  data,
  exposure,
  outcome,
  outcome_date = paste0(outcome, "_date"),
  start_date = "study_date",
  time_in = NULL,
  time_out = NULL,
  adjust_for = NULL,
  exclude = NULL,
  exposure_term = exposure,
  cluster = NULL,
  init_values = c(0),
  additional = "",
  return = c("model", "reduced-model", "coefs", "exposure"),
  return_fn = identity,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{exposure}{Name of column containing exposure variable.}

\item{outcome}{Name of the column containing status indicator
(normally 0 = no disease, 1 = disease).}

\item{outcome_date}{Name of column containing date of event or censoring
date. (Default: \verb{paste0(outcome, "_date)})}

\item{start_date}{Name of column containing date of start of follow up.
(Default: "study_date")}

\item{time_in}{Name of column containing starting time for interval data.}

\item{time_out}{Name of column containing follow up time (or ending time for
interval data).}

\item{adjust_for}{Character vector of terms to include in the model formula.}

\item{exclude}{Names of columns that identify observations that should be
excluded.}

\item{exposure_term}{The term(s) to include in the model for the exposure.
This can be a character vector or a function that take the exposure name as
the argument and returns a character vector. (Default: exposure)}

\item{cluster}{Name of column that clusters the observations, for the
purposes of a robust variance.}

\item{init_values}{A named vector of initial values for model coefficients.}

\item{additional}{Extra results to include.
See "Additional results" for details.}

\item{return}{What should the function return:
\itemize{
\item "model" (default): The model object (see ?survival::coxph.object).
\item "reduced-model": The model object with some large components removed.
\item "coefs": A data frame of all estimated coefficients and other key
model statistics.
\item "exposure": A data frame of estimated coefficient(s) for the exposure
and other key model statistics.
}}

\item{return_fn}{If return="model", name of function to apply to model before it's returned.
(Default: identity)}

\item{...}{Arguments passed to \code{survival::coxph()}}
}
\value{
Either a model object, reduced model object or data frame of
coefficients.
}
\description{
This function is used to estimate the (adjusted) association
between an exposure variable and an incident disease endpoint
using a Cox proportional hazards model.
It is a wrapper for the \link[survival]{coxph} function for convenience
and provides results from some additional analysis steps if requested.
}
\section{Time scale}{
For analysis of incident disease using time-on-study as the time scale,
\code{outcome}, \code{outcome_date} and \code{start_date} are used to define follow up
time (right censored data).

For other anlyses, the \code{time_in} and \code{time_out} arguments can be used.
If \code{time_in} and \code{time_out} are given, then these are starting and ending
time for interval survival data.
If \code{time_out} (but not \code{time_in}) is given, then this is the follow up time
(right censored data).
}

\section{Additional results}{
The \code{additional} argument is used to add results from additional analyses.
It must be a character vector containing one or more of:
\itemize{
\item "lrt": Likehood ratio test comparing the full model to one without the
exposure. Output will have additional columns "Chisq" (chi-square test
statistics), "Df" (degrees of freedom), "p_Chisq" (p-value).
\item "float": Standard errors for the exposure calculated by \link{float}.
Output will have additional columns "float_estimate", "float_se" and
"limits".
\item "qvcalc": Standard errors for the exposure calculated by \link{qvcalc}.
Output will have additional columns "qvcalc_estimate", "quasi_se" and
"worstErrors".
\item "" (default): Nothing.
}
These will only be included when relevant e.g. standard errors from
\link{float} when the exposure has more than two levels.
This only has an effect if \code{return} is "coefs" or "exposure".

See \link{float} and \link{qvcalc} for details about these results. Note that
"limits" and "worstErrors" do not correspond to particular groups but
accuracy other all contrasts.
}

\examples{
cox_model(ckbtools_participant_data,
          exposure = "bmi_calc",
          outcome = "all_cause_mortality",
          adjust_for = c("age_at_study_date", "strata(sex)"),
          return = "coefs")

cox_model(ckbtools_participant_data,
          exposure = "bmi_grp",
          outcome = "all_cause_mortality",
          adjust_for = c("age_at_study_date", "strata(sex)"),
          return = "coefs")

}
