% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-age-at-risk.R
\name{expand_age_at_risk}
\alias{expand_age_at_risk}
\title{Lexis expansion of data by age-at-risk}
\usage{
expand_age_at_risk(
  data,
  col.id = "csid",
  col.dob = "dob_anon",
  col.entry_date = "study_date",
  col.exit_date = "endpoint_date",
  col.status = "endpoint",
  col.keep = NULL,
  ages = c(0, 100),
  age.labels = paste0(ages[-length(ages)], "-", ages[-1]),
  null.status = 0
)
}
\arguments{
\item{data}{A data frame}

\item{col.id}{Name of column containing unique participant IDs}

\item{col.dob}{Name of column containing date of birth (Must be of class \code{Date})}

\item{col.entry_date}{Name of column containing entry date (Must be of class \code{Date})}

\item{col.exit_date}{Name of column containing exit date (Must be of class \code{Date})}

\item{col.status}{Name of column containing status}

\item{col.keep}{Names of additional columns to include in the result (e.g. exposure and covariate variables)}

\item{ages}{Vector of ages (years) to create age-at-risk strata. Any participant follow-up outside these ages is dropped. (Must be a numeric vector with distinct, increasing values. Default: c(0, 100))}

\item{age.labels}{Vector of name for age-at-risk strata (Default: \code{paste0(ages[-length(ages)], "-", ages[-1])})}

\item{null.status}{Value to use in status column for intervals before the final interval (Default: 0)}
}
\value{
Data frame expanded by age-at-risk
}
\description{
Lexis expansion of data by age-at-risk
}
