% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models-utils.R
\name{find_exposure_terms}
\alias{find_exposure_terms}
\title{Find terms in a model that relate to exposure}
\usage{
find_exposure_terms(data, model, exposure, exposure_term)
}
\description{
Find terms in a model that relate to exposure
}
\keyword{internal}
