% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models-utils.R
\name{get_qvcalc}
\alias{get_qvcalc}
\title{qvcalc::qvcalc::float as a tibble}
\usage{
get_qvcalc(model, exposure, exposure_terms)
}
\description{
qvcalc::qvcalc::float as a tibble
}
\keyword{internal}
