% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend-heterogeneity.R
\name{heterogeneity}
\alias{heterogeneity}
\title{Test for heterogeneity}
\usage{
heterogeneity(beta, se)
}
\arguments{
\item{beta}{vector of independent estimates}

\item{se}{vector of their standard errors}
}
\value{
A named list:
\itemize{
\item test statistic (chi-squared test statistic)
\item degrees of freedom (number of degrees of freedom used for the test, equal to \code{length(beta)-1})
\item p (p-value)
}
}
\description{
A chi-squared test (on \eqn{k - 1} degrees of freedom) for heterogeneity given a vector of \eqn{k} independent estimates and their standard errors.
}
\details{
See \code{vignette("trend_heterogeneity")} for details.
}
