% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence-rates.R
\name{incidence_rates}
\alias{incidence_rates}
\title{Calculate standardised incidence rates}
\usage{
incidence_rates(
  data,
  col.dob = "dob_anon",
  col.entry_date = "study_date",
  col.exit_date = "endpoint_date",
  col.status = "endpoint",
  ages = c(0, 100),
  group = NULL,
  adjust_for = NULL,
  per_pyar = 1000,
  weights = NULL,
  method = c("norm", "gamma")
)
}
\arguments{
\item{data}{A data frame.}

\item{col.dob}{Name of column containing date of birth. (Must be of class \code{Date}.)}

\item{col.entry_date}{Name of column containing entry date. (Must be of class \code{Date}.)}

\item{col.exit_date}{Name of column containing exit date. (Must be of class \code{Date}.)}

\item{col.status}{Name of column containing status. This should be 1 if the event has occurred at the exit date, or 0 if it has not.}

\item{ages}{Vector of ages (years) to create age-at-risk strata. (Must be a numeric vector with distinct, increasing values. Default: \code{c(0, 100)})}

\item{group}{Name of column containing groups to separately calculate incidence rates.}

\item{adjust_for}{Vector of names of columns containing adjustment variables. Include "agegrp" to adjust for age-at-risk strata.}

\item{per_pyar}{Crude and standardised incidence rates are given per \code{per_pyar} person-years at risk. (Default: 1000)}

\item{weights}{A data frame with columns that match the \code{adjust_for} variables, and one row for each combination of levels of those variables, and a column named ‘std_pyar’ with the weights (person-years at risk for that stratum in the reference/standard population).}

\item{method}{Method to calculate confidence intervals, one of "norm" or "gamma". Default "norm" is Normal approximation.}
}
\value{
A named list:
\itemize{
\item incidence_rates_in_all_strata
\item standardised_incidence_rates
}
}
\description{
Calculate standardised incidence rates
}
\section{Methods}{
See \code{vignette("incidence_rates")} for details.
}

