% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models.R
\name{logistic_model}
\alias{logistic_model}
\title{Estimate the association between an exposure and a binary outcome}
\usage{
logistic_model(
  data,
  exposure,
  outcome,
  adjust_for = NULL,
  exclude = NULL,
  exclude_controls = NULL,
  exposure_term = exposure,
  additional = "",
  return = c("model", "reduced-model", "coefs", "exposure"),
  return_fn = identity,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{exposure}{Name of column containing exposure variable.}

\item{outcome}{Name of the column containing binary outcome variable.}

\item{adjust_for}{Character vector of terms to include in the model formula.}

\item{exclude}{Names of columns that identify observations that are to be
excluded.}

\item{exclude_controls}{Names of columns that identify observations that are
to be exluded when outcome is equal to zero.}

\item{exposure_term}{The term(s) to include in the model for the exposure.
This can be a character vector or a function that take the exposure name as
the argument and returns a character vector. (Default: exposure)}

\item{additional}{Extra results to include.
See "Additional results" for details.}

\item{return}{What should the function return:
\itemize{
\item "model" (default): The model object (object of class "glm").
\item "reduced-model": The model object with some large components removed.
\item "coefs": A data frame of all estimated coefficients and other key
model statistics.
\item "exposure": A data frame of estimated coefficient(s) for the exposure
and other key model statistics.
}}

\item{return_fn}{If return="model", name of function to apply to model before it's returned.
(Default: identity)}

\item{...}{Arguments passed to \code{stats::glm()}}
}
\value{
Either a model object, reduced model object or data frame of
coefficients.
}
\description{
This function is used to estimate the (adjusted) association (log odds ratio)
between an exposure variable and a binary outcome variable using
a logistic regression model.
It is a wrapper for the \link[stats]{glm} function with \code{family = "binomial"}
for convenience and provides results from some additional analysis steps if
requested.
}
\section{Additional results}{
The \code{additional} argument is used to add results from additional analyses.
It must be a character vector containing one or more of:
\itemize{
\item "lrt": Likehood ratio test comparing the full model to one without the
exposure. Output will have additional columns "Chisq" (chi-square test
statistics), "Df" (degrees of freedom), "p_Chisq" (p-value).
\item "float": Standard errors for the exposure calculated by \link{float}
or by \link[Epi]{ftrend} if the model has no intercept term.
Output will have additional columns "float_estimate", "float_se" and
"limits".
\item "qvcalc": Standard errors for the exposure calculated by \link{qvcalc}.
Output will have additional columns "qvcalc_estimate", "quasi_se" and
"worstErrors".
\item "" (default): Nothing.
}
These will only be included when relevant e.g. standard errors from
\link{float} when the exposure has more than two levels.
This only has an effect if \code{return} is "coefs" or "exposure".

See \link{float} and \link{qvcalc} for details about these results. Note that
"limits" and "worstErrors" do not correspond to particular groups but
accuracy other all contrasts.
}

\examples{
logistic_model(ckbtools_participant_data,
               exposure = "bmi_calc",
               outcome = "has_diabetes",
               adjust_for = c("age_at_study_date", "sex"),
               return = "coefs")

logistic_model(ckbtools_participant_data,
               exposure = "bmi_grp",
               outcome = "has_diabetes",
               adjust_for = c("age_at_study_date", "sex"),
               return = "coefs")

}
