% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models.R
\name{multiple_models}
\alias{multiple_models}
\title{Multiple linear, logistic or Cox proportional hazards regression models}
\usage{
multiple_models(
  data,
  model,
  exposures,
  outcomes,
  outcome_dates = paste0(outcomes, "_date"),
  start_date = "study_date",
  time_in = NULL,
  time_out = NULL,
  age_at_risk = NULL,
  by = NULL,
  exclude = NULL,
  exclude_controls = NULL,
  exclude_by_exposure = NULL,
  exclude_by_outcome = NULL,
  exclude_controls_by_outcome = NULL,
  adjust_for = list(none = NULL),
  remove_adjust_for = NULL,
  add_adjust_for = NULL,
  cluster = NULL,
  init_values = NULL,
  return = c("coefs", "exposure", "reduced-model", "model"),
  verbose = if (rlang::is_interactive()) "progress" else "updates",
  dryrun = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame.}

\item{model}{Type of model to use:
\itemize{
\item "linear": Linear regression using \code{linear_model_safely()}.
\item "logistic": Logistic regression using \code{logistic_model_safely()}.
\item "cox": Cox proportional hazards model using \code{cox_model_safely()}.
}}

\item{exposures}{Names of columns containing exposure variables.}

\item{outcomes}{Names of columns containing outcome variables.}

\item{outcome_dates}{Names of columns containing outcome dates (for Cox model only).}

\item{start_date}{Name of column containing date of start of follow up.
(Default: "study_date")}

\item{time_in}{Names of columns containing starting time for interval data (for Cox model only).}

\item{time_out}{Names of columns containing follow up time (or ending time for
interval data) (for Cox model only).}

\item{age_at_risk}{A list of arguments for \code{age_at_risk()} to
expand data by age-at-risk for each outcome before fitting models for that
outcome. If this is used, "strata(agegrp)" is automatically added to each
element of \code{adjust_for}.}

\item{by}{Name of column (or multiple columns) to create subgroups.}

\item{exclude}{Names of columns that identify observations that should be
excluded.}

\item{exclude_controls}{Names of columns that identify observations that are
to be exluded when outcome is equal to zero.}

\item{exclude_by_exposure}{A named list of character vectors of column names.
When the name of an element matches the exposure, the column names will be
added to \code{exclude}.}

\item{exclude_by_outcome}{A named list of character vectors of column names.
When the name of an element matches the outcome, the column names will be
added to \code{exclude}.}

\item{exclude_controls_by_outcome}{A named list of character vectors of
column names. When the name of an element matches the outcome, the column
names will be added to \code{exclude_controls}.}

\item{adjust_for}{A named list of character vectors of terms to include in
model formulas.}

\item{remove_adjust_for}{A named list of character vectors of terms.
When the name of an element matches either the outcome or exposure, terms
given in the character vector will be removed from every element of
\code{adjust_for}.}

\item{add_adjust_for}{A named list of character vectors of terms.
When the name of an element matches either the outcome or exposure, terms
given in the character vector will be added to every element of \code{adjust_for}.}

\item{cluster}{Name of column that clusters the observations, for the
purposes of a robust variance.}

\item{init_values}{For Cox models, initial values for model coefficients.
This must be a data frame with columns "outcome", "subgroup", "adjustment",
"term" and "estimate", where term and estimate are the coefficient names and
initial values.}

\item{return}{What should the model function return:
\itemize{
\item "exposure" (default): A data frame of estimated coefficient(s) for the exposure
and other key model statistics.
\item "coefs": A data frame of all estimated coefficients and other key
model statistics.
\item "model": The model object.
\item "reduced-model": The model object with some large components removed.
}
Note that R model objects can be very large (they can contain the model data,
residuals, fitted values, etc.) so it is not recommended to set
\code{return="model"} for a large number of models.}

\item{verbose}{"progress" (default) to show progress bars, "updates" for
messages without progress bars, or "quiet" to hide.}

\item{dryrun}{Return only the data frame of models to be run.
Columns (except 'subgroup', 'adjustment' and 'outcome_name') are the
arguments that will be used in the model function. (Default: FALSE)}

\item{...}{Arguments passed to the model function.}
}
\value{
A "tibble" data frame. If \code{return} is "model" or "reduced-model"
then it will contain a list-column called "model".
}
\description{
Multiple linear, logistic or Cox proportional hazards regression models
}
\section{Iteration}{


This function provides a convenient way to iterate over models using
\code{\link[=linear_model]{linear_model()}}, \code{\link[=logistic_model]{logistic_model()}} or \code{\link[=cox_model]{cox_model()}}. First a data frame
defining all models is created. If the number of exposures is greater than or
equal to the number of outcomes (or Cox models with age-at-risk expansion are
used) then the models are split by outcome, otherwise they are split by
exposure. Within each set of models, \code{\link[purrr:pmap]{purrr::pmap()}} is used to iterate over
the fitting of models. Results are then combined into a single data frame.

For iteration of other models or analyses, you must write your own code. The
Iteration chapter of the first edition of “R for Data Science” provides an
introduction to iteration using R: \url{https://r4ds.had.co.nz/iteration.html}
}

\examples{

# linear regression models
multiple_models(
  ckbtools_participant_data,
  model = "linear",
  exposures = c("age_at_study_date", "sex", "bmi_calc", "bmi_grp"),
  outcomes = c("sbp_mean")
)

# logistic regression models
multiple_models(
  ckbtools_participant_data,
  model = "logistic",
  exposures = c("age_at_study_date", "sex", "bmi_calc", "bmi_grp"),
  outcomes = c("has_diabetes")
)

# Cox proportional hazards models
multiple_models(
  ckbtools_participant_data,
  model = "cox",
  exposures = c("age_at_study_date", "sex", "bmi_calc", "bmi_grp"),
  outcomes = c("all_cause_mortality", "ihd")
)


}
\seealso{
\code{\link[=linear_model]{linear_model()}}, \code{\link[=logistic_model]{logistic_model()}}, \code{\link[=cox_model]{cox_model()}},
\code{\link[=linear_model_safely]{linear_model_safely()}}, \code{\link[=logistic_model_safely]{logistic_model_safely()}}, \code{\link[=cox_model_safely]{cox_model_safely()}}
}
