% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models-utils.R
\name{organise_coefs}
\alias{organise_coefs}
\title{organise coefs data frame
either keep only exposure terms, or move them to the top}
\usage{
organise_coefs(return, coefs, exposure_terms)
}
\description{
organise coefs data frame
either keep only exposure terms, or move them to the top
}
\keyword{internal}
