% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics-preparation.R
\name{prepare_olink_data}
\alias{prepare_olink_data}
\title{Preparation of Olink protein data}
\usage{
prepare_olink_data(
  data,
  meta_data,
  transform = c("none"),
  qc = c("default", "none"),
  columns = NULL
)
}
\arguments{
\item{data}{Raw Olink protein data (imported from
data_baseline_olink_explore.rds or data_baseline_olink_explore.csv)}

\item{meta_data}{Olink meta data produced by \code{\link[=prepare_olink_meta_data]{prepare_olink_meta_data()}}.}

\item{transform}{Transformation to apply to protein values.
Currently only "none" is available.}

\item{qc}{QC steps to apply:
\itemize{
\item "default" (default): replace protein values with a QC warning with
a missing value
\item "none"
}}

\item{columns}{Other columns in raw data to include in output. For example,
set to "lod" to include the limit of detection for each protein.}
}
\description{
Preparation of Olink protein data
}
