% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics-preparation.R
\name{prepare_olink_meta_data}
\alias{prepare_olink_meta_data}
\title{Preparation of Olink meta data}
\usage{
prepare_olink_meta_data(
  data,
  raw_olink_protein_data,
  names = c("protein", "id"),
  columns = c("uniprot", "olinkid")
)
}
\arguments{
\item{data}{Raw CKB Olink meta data (imported from olink_explore_meta.rds or
olink_explore_meta.csv)}

\item{raw_olink_protein_data}{Raw Olink protein data (imported from
data_baseline_olink_explore.rds or data_baseline_olink_explore.csv)}

\item{names}{Naming scheme to use for proteins: "protein" (default) to use
assay names or "id" to use Olink IDs.}

\item{columns}{Optional columns to include in the meta data.
(Default: c("uniprot", "olinkid"))}
}
\description{
Preparation of Olink meta data
}
