% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics-preparation.R
\name{prepare_proteomics_analysis_dataset}
\alias{prepare_proteomics_analysis_dataset}
\title{Preparation of a proteomics analysis data set}
\usage{
prepare_proteomics_analysis_dataset(
  participant_data,
  protein_data,
  protein_meta_data,
  plate_correction = c("none", "median-subcohort", "mean-subcohort"),
  scaling = c("none", "center-subcohort", "scale-subcohort", "center", "scale", "rint"),
  id = "csid"
)
}
\arguments{
\item{participant_data}{A data frame of participant data}

\item{protein_data}{Protein data produced by \code{\link[=prepare_olink_data]{prepare_olink_data()}} or
\code{\link[=prepare_somascan_data]{prepare_somascan_data()}}.}

\item{protein_meta_data}{Corresponding protein meta data produed by
\code{\link[=prepare_olink_meta_data]{prepare_olink_meta_data()}} or \code{\link[=prepare_somascan_meta_data]{prepare_somascan_meta_data()}}.}

\item{plate_correction}{Type of plate correction to apply:
\itemize{
\item "none" (default)
\item "median-subcohort": For each protein, subtract the plate median (of
subcohort values) and add the overall median (of subcohort values).
\item "mean-subcohort": For each protein, subtract the plate mean (of
subcohort values) and add the overall mean (of subcohort values).
}}

\item{scaling}{Scaling to apply to values for each protein:
\itemize{
\item "none" (default)
\item "center-subcohort": Subtract the overall mean of subcohort values.
\item "scale-subcohort": Subtract the overall mean and divide by the sample
standard deviation of subcohort values.
\item "center": Subtract the overall mean.
\item "scale": Subtract the overall mean and divide by the sample standard
deviation.
\item "rint": Rank inverse normal transformation.
}}

\item{id}{Name of column used to join participant_data and protein_data.
(Default: "csid")}
}
\description{
Preparation of a proteomics analysis data set
}
