% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics-preparation.R
\name{prepare_somascan_data}
\alias{prepare_somascan_data}
\title{Prepare SomaScan protein data}
\usage{
prepare_somascan_data(
  data,
  sample_data,
  meta_data,
  transform = c("log2", "log", "log10", "none"),
  qc = c("default", "flagged", "none"),
  repeat_samples = c("remove", "retain"),
  columns = NULL
)
}
\arguments{
\item{data}{Raw CKB SomaScan protein data (imported from
data_baseline_somalogic.rds or data_baseline_somalogic.csv)}

\item{sample_data}{Raw CKB SomaScan protein sample information data (imported
from data_baseline_somalogic_samples.rds or
data_baseline_somalogic_samples.csv)}

\item{meta_data}{SomaScan meta data produced by
\code{\link[=prepare_somascan_meta_data]{prepare_somascan_meta_data()}}.}

\item{transform}{Transformation to apply to protein values:
\itemize{
\item "log2" (default): log base 2
\item "log10": log base 10
\item "none"
}}

\item{qc}{QC steps to apply.
\itemize{
\item "default" (default): Exclude participant samples with
hybridization control scale factor out of range.
\item "flagged": Exclude participants samples flagged for any normalization
acceptance criteria.
\item "none": Do not exclude any participant samples.
}}

\item{repeat_samples}{Default "remove" will remove data for the second sample
for participants who had a sample run twice. Use "retain" to keep this data;
the second samples will have sample_source=1.}

\item{columns}{Names of any additional columns from sample_data to include in
the output.}
}
\description{
This functions prepares a data set of SomaScan protein data
given the raw CKB SomaScan data.
}
\section{QC acceptance criteria}{
Hybridization control scale factor (\code{hybcontrolnormscale}) and median signal
normalization scale factors (\code{normscale_20}, \code{normscale_0_5},
\code{normscale_0_005}) have an acceptance range of 0.4-2.5.

Percentage of measurements used for ANML scale factor calculation
(\code{anmlfractionused_20}, \code{anmlfractionused_0_5}, \code{anmlfractionused_0_005}) has
a criteria of > 30\% (0.3).

Samples that meet these criteria are given a PASS for normalization
acceptance criteria for all row scale factors (\code{rowcheck}), while samples
that are outside the criteria are given a FLAG.

Hybridization control scale factor out of range suggests a technical issue.
Set \code{qc = "default"} to exclude participant samples with hybridization
control scale factor (\code{hybcontrolnormscale}) out of range (i.e. less than 0.4
or greater than 2.5).

Other criteria out of range suggests that the sample is fairly different from
a typical population.
Set \code{qc = "flagged"} to exclude participant samples with a FLAG for
normalization acceptance criteria for all row scale factors (\code{rowcheck}).
}

