% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics-preparation.R
\name{prepare_somascan_meta_data}
\alias{prepare_somascan_meta_data}
\title{Preparation of SomaScan meta data}
\usage{
prepare_somascan_meta_data(
  data,
  names = c("protein", "id"),
  columns = c("seqid", "somaid", "target_full_name", "target", "uniprot", "organism",
    "type", "dilution"),
  filter = "HumanProtein"
)
}
\arguments{
\item{data}{Raw CKB SomaScan meta data (imported from somalogic_meta.rds or
somalogic_meta.csv)}

\item{names}{Naming scheme to use for proteins: "protein" (default) to use
target name or "id" to use SomaLogic sequence IDs.}

\item{columns}{Optional columns to include in the meta data.
(Default: c("seqid", "somaid", "target_full_name", "target", "uniprot",
"organism", "type", "dilution"))}

\item{filter}{Filter to apply to rows:
\itemize{
\item "HumanProtein" (default): Only return meta data for aptamers where
organism is "Human" and type is "Protein".
\item "All": Return all rows of meta data.
}}
}
\description{
Preparation of SomaScan meta data
}
