% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models-utils.R
\name{reduce_model_object}
\alias{reduce_model_object}
\title{Reduce the size of model objects}
\usage{
reduce_model_object(x)
}
\description{
Reduce the size of model objects
}
\keyword{internal}
