% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{float}
\alias{qvcalc}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{Epi}{\code{\link[Epi:float]{float()}}}

  \item{qvcalc}{\code{\link[qvcalc:qvcalc]{qvcalc()}}}
}}

