% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models.R
\name{safemodels}
\alias{safemodels}
\alias{linear_model_safely}
\alias{logistic_model_safely}
\alias{cox_model_safely}
\title{Safe versions of regression model function}
\usage{
linear_model_safely(...)

logistic_model_safely(...)

cox_model_safely(...)
}
\arguments{
\item{...}{Arguments are passed to the model function.}
}
\value{
A named list:
\itemize{
\item result: What is returned by the model function.
\item errors: Errors produced by the model function as a character string.
\item warnings: Warnings produced by the model function as a character
string.
\item messages: Messages produced by the model function as a character
string.
\item output: Output produced by the model function as a character string.
}
}
\description{
These functions are wrappers around the regression model functions that
will capture side-effects such as errors and warnings and return a list.
This is particularly useful when using the model functions iteratively.
}
\seealso{
\code{\link[=linear_model]{linear_model()}}, \code{\link[=logistic_model]{logistic_model()}}, \code{\link[=cox_model]{cox_model()}}
}
