% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-data.R
\name{summarise_continuous}
\alias{summarise_continuous}
\title{Summarise continuous variables}
\usage{
summarise_continuous(data, cols, by = NULL)
}
\arguments{
\item{data}{Data set
(e.g. produced by \code{prepare_proteomics_analysis_dataset()})}

\item{cols}{Names of columns containing continuous variables to be
summarised.}

\item{by}{Names of columns that contain categorical participant
characteristics to produce subgroup summaries.}
}
\description{
Summarise continuous variables
}
