% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression-models-utils.R
\name{to_character_vector}
\alias{to_character_vector}
\title{Convert error/warning objects to character}
\usage{
to_character_vector(obj)
}
\description{
Convert error/warning objects to character
}
\keyword{internal}
