% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend-heterogeneity.R
\name{trend}
\alias{trend}
\title{Test for trend}
\usage{
trend(beta, se)
}
\arguments{
\item{beta}{vector of independent estimates}

\item{se}{vector of their standard errors}
}
\value{
A named list:
\itemize{
\item test statistic (chi-squared test statistic)
\item p (p-value)
}
}
\description{
A chi-squared test for trend (on 1 degree of freedom) given a vector of independent estimates and their standard errors.
}
\details{
See \code{vignette("trend_heterogeneity")} for details.
}
